refactor(backdown) remove gql code, stick to rest - As for the complexity this adds for UX and DX, I probably think it might not be a good idea to keep both rest and gql code branching here. If gql is really required we might want this to be its own separate template, or explore further how to properly do that with cookie-cutter
