+++ {{cookiecutter.project_slug}}/clients/react/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@apollo/client": "^3.4.17",
     "@chakra-ui/react": "2.3.4",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
@@ -13,8 +12,6 @@
     "axios": "^1.2.6",
     "cypress": "^8.6.0",
     "framer-motion": "^7.6.15",
-    "graphql": "^16.0.1",
-    "graphql-tag": "^2.12.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^3.0.0",
