+++ {{cookiecutter.project_slug}}/clients/react/src/app.tsx
@@ -1,8 +1,8 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import { BrowserRouter as Router, Outlet } from 'react-router-dom'
 import theme from './theme'
-import { AuthProvider } from './utils/auth.rest'
-import { AppRoutes } from './utils/routes.rest'
+import { AuthProvider } from './utils/auth'
+import { AppRoutes } from './utils/routes'
 
 export const AppRoot = () => {
   return (
