--- {{cookiecutter.project_slug}}/clients/react/src/main.rest.tsx
@@ -1,18 +1,0 @@
-import { ColorModeScript } from '@chakra-ui/react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import * as React from 'react'
-import { createRoot } from 'react-dom/client'
-import { App } from './app.rest'
-
-const container = document.getElementById('root')
-const root = createRoot(container!)
-const qClient = new QueryClient()
-
-root.render(
-  <React.StrictMode>
-    <ColorModeScript />
-    <QueryClientProvider client={qClient}>
-      <App />
-    </QueryClientProvider>
-  </React.StrictMode>,
-)
