+++ {{cookiecutter.project_slug}}/clients/react/src/utils/auth.tsx
