--- {{cookiecutter.project_slug}}/clients/react/src/utils/get-cookie.js
@@ -1,23 +1,0 @@
-/*
- * getCookie
- *
- * Use this method for getting cookie values (i.e. for getting the CSRF token).
- *
- * Usage:
- *   const csrfToken = getCookie("csrftoken");
- */
-export default function getCookie(cname) {
-  var name = cname + '='
-  var decodedCookie = decodeURIComponent(document.cookie)
-  var ca = decodedCookie.split(';')
-  for (var i = 0; i < ca.length; i++) {
-    var c = ca[i]
-    while (c.charAt(0) === ' ') {
-      c = c.substring(1)
-    }
-    if (c.indexOf(name) === 0) {
-      return c.substring(name.length, c.length)
-    }
-  }
-  return ''
-}
