--- {{cookiecutter.project_slug}}/clients/react/src/utils/mutations.ts
@@ -1,41 +1,0 @@
-import gql from 'graphql-tag'
-
-/* add GraphQL client-side mutations here */
-
-// Auth mutations
-export const LOG_IN = gql`
-  mutation logIn($email: String!, $password: String!) {
-    tokenAuth(email: $email, password: $password) {
-      token
-    }
-  }
-`
-
-export const REFRESH_TOKEN = gql`
-  mutation RefreshToken($token: String!) {
-    refreshToken(token: $token) {
-      token
-      payload
-      refreshExpiresIn
-    }
-  }
-`
-
-export const VERIFY_TOKEN = gql`
-  mutation VerifyToken($token: String!) {
-    verifyToken(token: $token) {
-      payload
-    }
-  }
-`
-
-export const CREATE_USER = gql`
-  mutation CreateUser($data: UserInput!) {
-    createUser(data: $data) {
-      user {
-        id
-        email
-      }
-    }
-  }
-`
