--- {{cookiecutter.project_slug}}/clients/react/src/utils/queries.ts
@@ -1,4 +1,0 @@
-/* import gql from "graphql-tag"; */
-
-/* add GraphQL client-side queries here */
-export const sampleGql = ``
