+++ {{cookiecutter.project_slug}}/clients/react/src/utils/routes.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { Route, Routes } from 'react-router-dom'
 import { Home, Layout } from 'src/pages'
 //TODO: move these to pages index so you import them from line above
-import { LogIn } from 'src/pages/log-in.gql'
-import { SignUp } from 'src/pages/sign-up.gql'
+import { LogIn } from 'src/pages/log-in'
+import { SignUp } from 'src/pages/sign-up'
 
 export const PUBLIC_ROUTES = ['/home', '/log-in', '/sign-up']
